+++ baybe/surrogates/gaussian_process/presets/__init__.py
@@ -1,5 +1,6 @@
 """Gaussian process surrogate presets."""
 
+from baybe.surrogates.gaussian_process.presets.botorch import BotorchKernelFactory
 from baybe.surrogates.gaussian_process.presets.core import (
     GaussianProcessPreset,
     make_gp_from_preset,
@@ -10,6 +11,7 @@ from baybe.surrogates.gaussian_process.presets.edbo import EDBOKernelFactory
 __all__ = [
     "DefaultKernelFactory",
     "EDBOKernelFactory",
+    "BotorchKernelFactory",
     "make_gp_from_preset",
     "GaussianProcessPreset",
 ]